feat: comment Entity package 이동 (#17)
