--- src/main/java/com/ssu/travel/comment/Comment.java
@@ -1,44 +1,0 @@
-package com.ssu.travel.comment;
-
-import com.ssu.travel.post.Post;
-import com.ssu.travel.user.User;
-import lombok.*;
-
-import javax.persistence.*;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @Column(columnDefinition = "TEXT")
-    private String content;
-
-    public void updateContent(String content) {
-        this.content = content;
-    }
-
-
-    private Comment(Post post, User user, String content) {
-        this.post = post;
-        this.user = user;
-        this.content = content;
-    }
-
-    public static Comment of(Post post, User user, String content) {
-        return new Comment(post, user, content);
-    }
-}
